requirements.txt
selenium
webdriver-manager
tqdm
pillow
pyautogui
selenium-wire
img2pdf